Python/Google_Meet_Scheduler/requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pathlib
pickleshare
typing
uuid